Keep tree distances whole instead of truncating them to integers
"Keep widths and heights of flaps" could fail with "Optimizer result
violates distance N between flaps A and B" — the solver reporting
success and then its own checker rejecting the layout it had just
produced.

The kernel narrowed the distance map to i32. Every use immediately
widened it back with f64::from, so the integer was only ever a lossy
container at the boundary, but the truncation made the solver enforce a
*shorter* separation than the tree requires. The checker reads the
untruncated distance from the request, so the two disagreed.

Distances are fractional exactly when flap dimensions are in play:
upstream's getArea divides by PI, and getDistMap takes the square root
of that as a leaf's length. For the reported design that is
sqrt(2/PI + 1) = 1.2793..., giving 3.2793... between flaps 5 and 8,
truncated to 3. With dimensions off every distance is a whole number and
the truncation is a no-op, which is why unchecking the box worked.

Upstream keeps these as plain numbers throughout (DistMap is
[T, T, number][]), so the integer was a porting error rather than a
quirk worth preserving.

Two reasons this survived: the in-crate conversion test asserted the
truncated value, encoding the bug as expected behaviour; and every case
in the oracle batch uses whole-number distances, so 410 differential
cases could pass with it present. The oracle still passes unchanged
(exact=209, valid_different=201, 0 invalid).

The regression test drives the reported design end to end. A synthetic
two-flap case is packed loosely enough to satisfy both the truncated and
the true distance by luck — it passed with the bug restored.

Includes the rebuilt wasm bridge; the tracked artifact is what the
browser runs.

+++ crates/oristudio-bp/examples/symmetry_showcase.rs
@@ -46,7 +46,7 @@ struct Case {
     note: String,
     sheet: OptimizerSheet,
     leaves: Vec<Leaf>,
-    dist: Vec<(usize, usize, i32)>,
+    dist: Vec<(usize, usize, f64)>,
     partners: Vec<(u32, u32)>,
 }
 
@@ -117,7 +117,7 @@ fn bug(vertebrae: usize, spine: i32, leg: i32, head: i32) -> Case {
     for i in 0..leaves.len() {
         for j in (i + 1)..leaves.len() {
             let along = (offset[i] - offset[j]).abs();
-            dist.push((i, j, leaves[i].length + along + leaves[j].length));
+            dist.push((i, j, f64::from(leaves[i].length + along + leaves[j].length)));
         }
     }
     Case {
@@ -150,7 +150,7 @@ fn star(n: usize, leg: i32) -> Case {
     let mut dist = Vec::new();
     for i in 0..n {
         for j in (i + 1)..n {
-            dist.push((i, j, 2 * leg));
+            dist.push((i, j, f64::from(2 * leg)));
         }
     }
     Case {
@@ -188,7 +188,7 @@ fn creature(segments: &[(i32, i32)], spine: i32, head: i32, tail: i32) -> Case {
     for i in 0..leaves.len() {
         for j in (i + 1)..leaves.len() {
             let along = (offset[i] - offset[j]).abs();
-            dist.push((i, j, leaves[i].length + along + leaves[j].length));
+            dist.push((i, j, f64::from(leaves[i].length + along + leaves[j].length)));
         }
     }
     Case {
@@ -212,7 +212,7 @@ fn sized_pair() -> Case {
     let mut dist = Vec::new();
     for i in 0..leaves.len() {
         for j in (i + 1)..leaves.len() {
-            dist.push((i, j, leaves[i].length + leaves[j].length));
+            dist.push((i, j, f64::from(leaves[i].length + leaves[j].length)));
         }
     }
     Case {
